Build 117 Changes
- Moved away from yarn to npm for package management (#32)
- Modified .gitignore to be more greedy w/Webpack statistics files.
- Cleaned-up and updated packages (#88). This includes a fix to the `visualReducer`, as the Power BI team have made a breaking change in API 3.8 for `allowInteractions`.
- Moved `event` API into `core/interactivity` (#91)
- Moved `tooltip` API into `core/interactivity/tooltip` (#90)
