+++ .gitignore
@@ -67,4 +67,4 @@ typings/
 .api/
 .tmp/
 dist/
-webpack.statistics.*.html
+webpack.statistics*.html
