+++ package.json
@@ -16,28 +16,29 @@
         "precommit-msg"
     ],
     "devDependencies": {
-        "@babel/polyfill": "^7.12.1",
-        "@babel/runtime": "7.14.0",
-        "@babel/runtime-corejs2": "7.14.0",
+        "@babel/core": "^7.15.0",
+        "@babel/plugin-transform-runtime": "^7.15.0",
+        "@babel/preset-env": "^7.15.0",
+        "@babel/runtime": "^7.14.8",
+        "@babel/runtime-corejs3": "^7.14.9",
         "@fluentui/react": "^8.14.12",
         "@reduxjs/toolkit": "^1.5.0",
         "@types/ace": "^0.0.44",
         "@types/d3-format": "^2.0.0",
-        "@types/d3-time-format": "^3.0.0",
         "@types/d3-selection": "^3.0.0",
+        "@types/d3-time-format": "^3.0.0",
         "@types/jest": "^26.0.23",
         "@types/jsoneditor": "^8.6.0",
         "@types/lodash": "^4.14.168",
+        "@types/node": "^16.4.12",
         "@types/react": "^17.0.0",
         "@types/react-dom": "^17.0.0",
         "@types/react-redux": "^7.1.15",
-        "@types/react-split-pane": "^0.1.67",
-        "@types/vega": "^3.2.0",
         "ace-builds": "^1.4.12",
         "adm-zip": "^0.5.5",
         "ajv": "^6.12.6",
         "axios": "^0.21.1",
-        "core-js": "3.6.0",
+        "core-js": "^3.16.0",
         "d3-format": "^2.0.0",
         "d3-selection": "^3.0.0",
         "d3-time-format": "^3.0.0",
@@ -45,7 +46,7 @@
         "jsoneditor": "^9.5.1",
         "jsonrepair": "^2.2.1",
         "lodash": "^4.17.21",
-        "powerbi-visuals-api": "^3.7.0",
+        "powerbi-visuals-api": "^3.8.2",
         "powerbi-visuals-utils-dataviewutils": "2.2.1",
         "powerbi-visuals-utils-formattingutils": "^4.7.1",
         "powerbi-visuals-utils-interactivityutils": "^5.7.0",
@@ -53,17 +54,16 @@
         "prettier": "^2.2.1",
         "process": "^0.11.10",
         "react": "^17.0.1",
-        "react-ace": "^9.2.1",
+        "react-ace": "^9.4.3",
         "react-dom": "^17.0.1",
         "react-hotkeys-hook": "^3.3.2",
-        "react-icons": "^4.1.0",
         "react-redux": "^7.2.2",
         "react-split-pane": "^0.1.92",
         "react-vega": "7.4.2",
         "redux": "^4.0.5",
         "regenerator-runtime": "^0.13.7",
         "ts-jest": "^26.5.6",
-        "ts-loader": "6.1.0",
+        "ts-loader": "^9.2.5",
         "tslint": "^5.18.0",
         "tslint-microsoft-contrib": "^6.2.0",
         "typescript": "^4.2.4",
