--- src/api/event/README.md
@@ -1,24 +1,0 @@
-# Deneb - Event API
-
-API for managing events within the visual.
-
--   [Constants](#constants)
--   [Methods](#methods)
--   [Interfaces](#interfaces)
--   [Types](#types)
-
-[< index](../README.md)
-
-## Constants
-
-## Methods
-
-#### `resolveCoordinates`(_event_)
-
-For the supplied event, returns an [x, y] array of mouse coordinates.
-
-## Private Methods
-
-## Interfaces
-
-## Types
