--- src/api/event/index.ts
@@ -1,6 +1,0 @@
-export { resolveCoordinates };
-
-const resolveCoordinates = (event: MouseEvent): [number, number] => [
-    event.clientX,
-    event.clientY
-];
