+++ src/api/index.ts
@@ -3,7 +3,6 @@ import * as dataset from './dataset';
 import * as dataView from './dataView';
 import * as developer from './developer';
 import * as editor from './editor';
-import * as event from './event';
 import * as features from './features';
 import * as fluent from './fluent';
 import * as formatting from './formatting';
@@ -12,7 +11,6 @@ import * as selection from './selection';
 import * as specification from './specification';
 import * as store from './store';
 import * as template from './template';
-import * as tooltip from './tooltip';
 import * as ui from './ui';
 
 export {
@@ -21,7 +19,6 @@ export {
     dataView,
     developer,
     editor,
-    event,
     fluent,
     features,
     formatting,
@@ -30,6 +27,5 @@ export {
     specification,
     store,
     template,
-    tooltip,
     ui
 };
